Finished implementing the read_data.py file
This commit finished the implementation of the read_data.py file.
The file is now able to read the data from the CSV file and return it as a list of dictionaries.
There is some example code included in main.py, as we are not using it yet. Replace it when needed.

* Renamed app.py to read_data.py
* Added the main.py file